Separate parse message methods from uploads
