+++ telethon/utils.py
@@ -12,6 +12,7 @@ import types
 from collections import UserList
 from mimetypes import guess_extension
 
+from .extensions import markdown, html
 from .tl import TLObject
 from .tl.types import (
     Channel, ChannelForbidden, Chat, ChatEmpty, ChatForbidden, ChatFull,
@@ -421,6 +422,39 @@ def get_message_id(message):
         pass
 
     raise TypeError('Invalid message type: {}'.format(type(message)))
+
+
+def sanitize_parse_mode(mode):
+    """
+    Converts the given parse mode into an object with
+    ``parse`` and ``unparse`` callable properties.
+    """
+    if not mode:
+        return None
+
+    if callable(mode):
+        class CustomMode:
+            @staticmethod
+            def unparse(text, entities):
+                raise NotImplementedError
+
+        CustomMode.parse = mode
+        return CustomMode
+    elif (all(hasattr(mode, x) for x in ('parse', 'unparse'))
+          and all(callable(x) for x in (mode.parse, mode.unparse))):
+        return mode
+    elif isinstance(mode, str):
+        try:
+            return {
+                'md': markdown,
+                'markdown': markdown,
+                'htm': html,
+                'html': html
+            }[mode.lower()]
+        except KeyError:
+            raise ValueError('Unknown parse mode {}'.format(mode))
+    else:
+        raise TypeError('Invalid parse mode type {}'.format(mode))
 
 
 def get_input_location(location):
